- improved attachment migration - only attachments that are not part of a message are automatically removed now when the file is gone - other attachments without files are marked with a new fileStatus “DOES_NOT_EXIST” - set files owned by message attachments to read-only and keep the WEBDAV-Server from deleting them - label fixes from songs->files

+++ HoccerXO/Audio Player/HXOAudioPlayer.m
@@ -182,7 +182,7 @@
     self.playlistIndex = index;
     
     Attachment *attachment = [self.playlist attachmentAtIndex:self.currentTrackNumber];
-    NSLog(@"playAtIndex %d, attachment = %@", index, attachment.humanReadableFileName);
+    NSLog(@"playAtIndex %lu, attachment = %@", (unsigned long)index, attachment.humanReadableFileName);
     return [self playAttachment:attachment];
 }
